+++ stac_fastapi/api/stac_fastapi/api/config.py
@@ -18,6 +18,7 @@ class ApiExtensions(enum.Enum):
     query = "query"
     sort = "sort"
     transaction = "transaction"
+    collection_search = "collection-search"
 
 
 class AddOns(enum.Enum):